qualcommax: nvmem: layouts fix mac base ascii cells handling
The ascii-type value passed from the ascii-env driver for the mac address
is 18 bytes of length instead of 17 including the trailing null character.
So, let's fix the validation to properly parse the mac address.

Tested on Linksys MX2000 and MX5500 on qualcommax target only.
Hence, created a platform specific patch.

Signed-off-by: George Moussalem <george.moussalem@outlook.com>